mem fix: use pthread_join() instead of pthread_detach()
The use of pthread_join() allows the master thread to wait until all
threads have exited successfully. It's probably better to do it this way
than detaching the threads and trying to manually verify if the thread
did exit completely. It's also a lot easier to do it this way, since all
we have to do is call pthread_join().

Also added comments on master thread's wait loop.