movegen: fix undo_info bit alignment bug
This was a serious bug that affected move generation based on zobrist
keys (hashtables). The reason was that the zobrist keys became corrupted
because undo_info was blindly copied into pos->info, when undo_info was
*NOT* aligned the same way as pos->info. Specifically, it was missing
the 6 bits that pos->info has to account for rook birthfiles. This
resulted in incorrect reads for the en passant square.

The fix is simple: make undo_info's critical bits aligned perfectly with
pos->info, to ensure that ep_sq_uinfo(), et al always reads the correct
bits, and that (pos->info |= undo_info) is always correct.